Added book creation, edit, delete tests to the suite
