--- app/src/androidTest/java/com/example/atheneum/activities/AddRemoveBookTest.java
@@ -1,86 +1,0 @@
-package com.example.atheneum.activities;
-
-import android.app.Activity;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.espresso.ViewInteraction;
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
-import android.view.View;
-import android.view.ViewGroup;
-import android.view.ViewParent;
-
-import com.android.dx.command.Main;
-import com.example.atheneum.R;
-import com.example.atheneum.utils.ConnectionChecker;
-import com.example.atheneum.utils.FirebaseAuthUtils;
-import com.robotium.solo.Solo;
-
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeMatcher;
-import org.hamcrest.core.IsInstanceOf;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static android.support.test.InstrumentationRegistry.getContext;
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static android.support.test.espresso.action.ViewActions.replaceText;
-import static android.support.test.espresso.action.ViewActions.scrollTo;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
-import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
-import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
-import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static junit.framework.TestCase.assertFalse;
-import static junit.framework.TestCase.assertTrue;
-import static org.hamcrest.Matchers.allOf;
-import static org.hamcrest.Matchers.is;
-
-@LargeTest
-@RunWith(AndroidJUnit4.class)
-public class AddRemoveBookTest extends ActivityTestRule<MainActivity>{
-    private Solo solo;
-
-    public AddRemoveBookTest() {
-        super(MainActivity.class, false, false);
-    }
-
-    @Rule
-    public ActivityTestRule<MainActivity> activityTestRule =
-            new ActivityTestRule<>(MainActivity.class, false, true);
-
-    @Before
-    public void setUp() throws Exception {
-        solo = new Solo(InstrumentationRegistry.getInstrumentation(),
-                activityTestRule.getActivity());
-    }
-
-
-    @Test
-    public void start() throws Exception{
-        Activity activity = activityTestRule.getActivity();
-    }
-
-
-    @Test
-    public void addRemoveBook() throws  Exception {
-        solo.assertCurrentActivity("Wrong Activity", MainActivity.class);
-
-        solo.clickOnActionBarItem(R.id.nav_view);
-        Thread.sleep(2000);
-
-    }
-
-    @After
-    public void tearDown() throws Exception{
-        solo.finishOpenedActivities();
-    }
-
-}
