+++ app/src/androidTest/java/com/example/atheneum/activities/ViewEditBookTest.java
@@ -1,0 +1,310 @@
+package com.example.atheneum.activities;
+
+import android.support.test.espresso.ViewInteraction;
+import android.support.test.filters.LargeTest;
+import android.support.test.rule.ActivityTestRule;
+import android.support.test.runner.AndroidJUnit4;
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+import android.view.ViewGroup;
+import android.view.ViewParent;
+
+import com.example.atheneum.R;
+import com.example.atheneum.utils.FirebaseAuthUtils;
+
+import org.hamcrest.Description;
+import org.hamcrest.Matcher;
+import org.hamcrest.TypeSafeMatcher;
+import org.hamcrest.core.IsInstanceOf;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.action.ViewActions.click;
+import static android.support.test.espresso.action.ViewActions.closeSoftKeyboard;
+import static android.support.test.espresso.action.ViewActions.replaceText;
+import static android.support.test.espresso.action.ViewActions.scrollTo;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
+import static android.support.test.espresso.contrib.RecyclerViewActions.actionOnItemAtPosition;
+import static android.support.test.espresso.matcher.ViewMatchers.Visibility.VISIBLE;
+import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
+import static android.support.test.espresso.matcher.ViewMatchers.withClassName;
+import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
+import static android.support.test.espresso.matcher.ViewMatchers.withEffectiveVisibility;
+import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static android.support.test.espresso.matcher.ViewMatchers.withText;
+import static junit.framework.TestCase.assertTrue;
+import static org.hamcrest.Matchers.allOf;
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.not;
+
+@LargeTest
+@RunWith(AndroidJUnit4.class)
+
+public class ViewEditBookTest {
+    private static final long DELAY_MS = 500;
+
+    @Rule
+    public ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
+
+    @Test
+    public void viewEditBookTest() {
+        assertTrue("Must be logged in before trying to view or edit a book!", FirebaseAuthUtils.isCurrentUserAuthenticated());
+
+        // Added a sleep statement to match the app's execution delay.
+        // The recommended way to handle such scenarios is to use Espresso idling resources:
+        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
+        try {
+            Thread.sleep(DELAY_MS);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        // Navigate to owner page
+        ViewInteraction appCompatImageButton = onView(
+                allOf(withContentDescription("Open navigation drawer"),
+                        childAtPosition(
+                                allOf(withId(R.id.toolbar),
+                                        childAtPosition(
+                                                withClassName(is("android.support.design.widget.AppBarLayout")),
+                                                0)),
+                                1),
+                        isDisplayed()));
+        appCompatImageButton.perform(click());
+
+        ViewInteraction navigationMenuItemView = onView(
+                allOf(childAtPosition(
+                        allOf(withId(R.id.design_navigation_view),
+                                childAtPosition(
+                                        withId(R.id.nav_view),
+                                        0)),
+                        3),
+                        isDisplayed()));
+        navigationMenuItemView.perform(click());
+
+        // click on existing book
+        ViewInteraction recyclerView = onView(
+                allOf(withId(R.id.owner_books_recycler_view),
+                        childAtPosition(
+                                withClassName(is("android.widget.LinearLayout")),
+                                0)));
+        recyclerView.perform(actionOnItemAtPosition(0, click()));
+
+        // Added a sleep statement to match the app's execution delay.
+        // The recommended way to handle such scenarios is to use Espresso idling resources:
+        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
+        try {
+            Thread.sleep(DELAY_MS);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        // validate the data in book info
+        ViewInteraction textView = onView(
+                allOf(withId(R.id.bookTitle), withText("Testbook"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                2),
+                        isDisplayed()));
+        textView.check(matches(withText("Testbook")));
+
+        ViewInteraction textView1 = onView(
+                allOf(withId(R.id.bookAuthor), withText("logintester"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                4),
+                        isDisplayed()));
+        textView1.check(matches(withText("logintester")));
+
+        ViewInteraction textView2 = onView(
+                allOf(withId(R.id.bookISBN), withText("111"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                6),
+                        isDisplayed()));
+        textView2.check(matches(withText("111")));
+
+        ViewInteraction textView3 = onView(
+                allOf(withId(R.id.bookStatus), withText("AVAILABLE"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                8),
+                        isDisplayed()));
+        textView3.check(matches(withText("AVAILABLE")));
+
+        ViewInteraction textView4 = onView(
+                allOf(withId(R.id.book_borrower_email),
+                        childAtPosition(
+                                allOf(withId(R.id.borrower_prof_area),
+                                        childAtPosition(
+                                                IsInstanceOf.<View>instanceOf(android.widget.LinearLayout.class),
+                                                10)),
+                                1),
+                        isDisplayed()));
+        textView4.check(matches(withText("None")));
+
+        ViewInteraction textView5 = onView(
+                allOf(withId(R.id.bookDescription), withText("Lorem ipsum"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                12),
+                        isDisplayed()));
+        textView5.check(matches(withText("Lorem ipsum")));
+
+        // edit the book and check the results
+        ViewInteraction appCompatButton3 = onView(
+                allOf(withId(R.id.buttonEdit), withText("Edit"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withClassName(is("android.widget.LinearLayout")),
+                                        15),
+                                1)));
+        appCompatButton3.perform(scrollTo(), click());
+
+        // Added a sleep statement to match the app's execution delay.
+        // The recommended way to handle such scenarios is to use Espresso idling resources:
+        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
+        try {
+            Thread.sleep(DELAY_MS);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        // edit the title
+        ViewInteraction appCompatEditText = onView(
+                allOf(withId(R.id.bookTitleEditText), withText("Testbook"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withClassName(is("android.widget.ScrollView")),
+                                        0),
+                                2)));
+        appCompatEditText.perform(scrollTo(), replaceText("New Title"));
+
+        ViewInteraction appCompatEditText1 = onView(
+                allOf(withId(R.id.bookTitleEditText), withText("New Title"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withClassName(is("android.widget.ScrollView")),
+                                        0),
+                                2),
+                        isDisplayed()));
+        appCompatEditText1.perform(closeSoftKeyboard());
+
+        // edit the description
+        ViewInteraction appCompatEditText2 = onView(
+                allOf(withId(R.id.descEditText), withText("Lorem ipsum"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withClassName(is("android.widget.ScrollView")),
+                                        0),
+                                10)));
+        appCompatEditText2.perform(scrollTo(), replaceText("Some words"));
+
+        ViewInteraction appCompatEditText3 = onView(
+                allOf(withId(R.id.descEditText), withText("Some words"),
+                        childAtPosition(
+                                childAtPosition(
+                                        withClassName(is("android.widget.ScrollView")),
+                                        0),
+                                10),
+                        isDisplayed()));
+        appCompatEditText3.perform(closeSoftKeyboard());
+
+        ViewInteraction floatingActionButton = onView(
+                allOf(withId(R.id.saveBookBtn),
+                        childAtPosition(
+                                childAtPosition(
+                                        withId(android.R.id.content),
+                                        0),
+                                1),
+                        isDisplayed()));
+        floatingActionButton.perform(click());
+
+        // Added a sleep statement to match the app's execution delay.
+        // The recommended way to handle such scenarios is to use Espresso idling resources:
+        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
+        try {
+            Thread.sleep(DELAY_MS);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        // check title change
+        ViewInteraction textView6 = onView(
+                allOf(withId(R.id.book_title), withEffectiveVisibility(VISIBLE), withText("New Title"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.LinearLayout.class),
+                                        0),
+                                0),
+                        isDisplayed()));
+        textView6.check(matches(withText("New Title")));
+
+        // open the book information again to validate description change
+        ViewInteraction recyclerView2 = onView(
+                allOf(withId(R.id.owner_books_recycler_view),
+                        childAtPosition(
+                                withClassName(is("android.widget.LinearLayout")),
+                                0)));
+        recyclerView2.perform(actionOnItemAtPosition(0, click()));
+
+        // Added a sleep statement to match the app's execution delay.
+        // The recommended way to handle such scenarios is to use Espresso idling resources:
+        // https://google.github.io/android-testing-support-library/docs/espresso/idling-resource/index.html
+        try {
+            Thread.sleep(DELAY_MS);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        ViewInteraction textView11 = onView(
+                allOf(withId(R.id.bookTitle), withText("New Title"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                2),
+                        isDisplayed()));
+        textView11.check(matches(withText("New Title")));
+
+        ViewInteraction textView12 = onView(
+                allOf(withId(R.id.bookDescription), withText("Some words"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(android.widget.ScrollView.class),
+                                        0),
+                                12),
+                        isDisplayed()));
+        textView12.check(matches(withText("Some words")));
+    }
+
+    private static Matcher<View> childAtPosition(
+            final Matcher<View> parentMatcher, final int position) {
+
+        return new TypeSafeMatcher<View>() {
+            @Override
+            public void describeTo(Description description) {
+                description.appendText("Child at position " + position + " in parent ");
+                parentMatcher.describeTo(description);
+            }
+
+            @Override
+            public boolean matchesSafely(View view) {
+                ViewParent parent = view.getParent();
+                return parent instanceof ViewGroup && parentMatcher.matches(parent)
+                        && view.equals(((ViewGroup) parent).getChildAt(position));
+            }
+        };
+    }
+}
